+++ source/platforms/github/_tests/_github_git.test.ts
@@ -64,7 +64,7 @@ describe("the dangerfile gitDSL", () => {
 
     nodeGitHubAPI = new NodeGitHub()
     const mockContents = async ({ ref }: any) => (await requestWithFixturedJSON(`static_file.${ref}.json`))()
-    nodeGitHubAPI.repos.getContents = mockContents as any
+    nodeGitHubAPI.repos.getContent = mockContents as any
 
     gitJSONDSL = await github.getPlatformGitRepresentation()
     const githubJSONDSL = await github.getPlatformReviewDSLRepresentation()